blyat

+++ core/src/com/mygdx/mass/Map.java
@@ -2,8 +2,8 @@ package com.mygdx.mass;
 
 public class Map {
 
-    public static final int WIDTH = 200;
-    public static final int HEIGHT = 200;
+    public static final double WIDTH = 200;
+    public static final double HEIGHT = 200;
